LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-media-expected.txt


FAIL Style rules within @media are scoped assert_equals: expected "1" but got "auto"
